+++ catatom2osm/__init__.py
@@ -1,2 +1,1 @@
 __version__ = '2.3.0'
-from catatom2osm.__main__ import run